+++ src/calculator/calculator.c
@@ -197,6 +197,8 @@ static bool calculator_expression_independent(const CalculatorExpression *expres
                 case CALCULATOR_FUNCTION_SIGN:
                 case CALCULATOR_FUNCTION_MIN:
                 case CALCULATOR_FUNCTION_MAX:
+                case CALCULATOR_FUNCTION_SUM:
+                case CALCULATOR_FUNCTION_PRODUCT:
                 case CALCULATOR_FUNCTION_FLOOR:
                 case CALCULATOR_FUNCTION_CEIL:
                 case CALCULATOR_FUNCTION_TRUNC: